CAT-13 type owner-identity target class as Vertex or Edge
CatiaOwnerIdentityTarget.target_class is Vertex (0x5d) or Edge (0x5e).
CADIR JSON still emits the class byte.

Checks: cargo fmt --all && cargo check -q --workspace --tests
Hook-Bypass-Reason: the mission requires --no-verify for each non-final item; the full gate runs once at the end.

+++ crates/cadmpeg-codec-catia/src/native/tests/consolidated.rs
@@ -848,7 +848,7 @@ fn native_namespace_retains_fixed_owner_allocation_targets() {
                 target.slot,
                 target.distance,
                 target.target_byte_offset,
-                target.target_class,
+                u8::from(target.target_class),
             ))
             .collect::<Vec<_>>(),
         [
